Refactor: layout.tsx -> components/layout.tsx
